+++ src/pc/network/packets/packet_reservation_use.c
@@ -10,7 +10,7 @@
 //#define DISABLE_MODULE_LOG 1
 #include "pc/debuglog.h"
 
-void network_send_reservation_use(u8 syncId) {
+void network_send_reservation_use(u32 syncId) {
     SOFT_ASSERT(gNetworkType == NT_CLIENT);
 
     // make sure this is a reserved id
@@ -20,12 +20,12 @@ void network_send_reservation_use(u8 syncId) {
     packet_init(&p, PACKET_RESERVATION_USE, true, PLMT_NONE);
 
     extern s16 gCurrCourseNum, gCurrActStarNum, gCurrLevelNum, gCurrAreaIndex;
-    packet_write(&p, &gCurrCourseNum,  sizeof(u8));
-    packet_write(&p, &gCurrActStarNum, sizeof(u8));
-    packet_write(&p, &gCurrLevelNum,   sizeof(u8));
-    packet_write(&p, &gCurrAreaIndex,  sizeof(u8));
+    packet_write(&p, &gCurrCourseNum,  sizeof(s16));
+    packet_write(&p, &gCurrActStarNum, sizeof(s16));
+    packet_write(&p, &gCurrLevelNum,   sizeof(s16));
+    packet_write(&p, &gCurrAreaIndex,  sizeof(s16));
 
-    packet_write(&p, &syncId, sizeof(u8));
+    packet_write(&p, &syncId, sizeof(u32));
 
     network_send_to(gNetworkPlayerServer->localIndex, &p);
     LOG_INFO("tx reservation use");
@@ -41,19 +41,19 @@ void network_receive_reservation_use(struct Packet* p) {
         return;
     }
 
-    u8 courseNum, actNum, levelNum, areaIndex;
-    packet_read(p, &courseNum, sizeof(u8));
-    packet_read(p, &actNum,    sizeof(u8));
-    packet_read(p, &levelNum,  sizeof(u8));
-    packet_read(p, &areaIndex, sizeof(u8));
+    s16 courseNum, actNum, levelNum, areaIndex;
+    packet_read(p, &courseNum, sizeof(s16));
+    packet_read(p, &actNum,    sizeof(s16));
+    packet_read(p, &levelNum,  sizeof(s16));
+    packet_read(p, &areaIndex, sizeof(s16));
 
     if (courseNum != np->currCourseNum || actNum != np->currActNum|| levelNum != np->currLevelNum || areaIndex != np->currAreaIndex) {
         LOG_ERROR("received an improper location");
         return;
     }
 
-    u8 syncId;
-    packet_read(p, &syncId, sizeof(u8));
+    u32 syncId;
+    packet_read(p, &syncId, sizeof(u32));
 
     reservation_area_use(np, syncId);
 }
